+++ src/components/Card.js
@@ -1,18 +1,19 @@
-import { openModal } from "../utils/utils.js";
+// import { openModal } from "../utils/utils.js";
+// import Popup from "./Popup.js";
 
-const previewImageModal = document.querySelector("#preview-image");
-const imgPreview = previewImageModal.querySelector(
-  ".modal__card-image-preview"
-);
-const titlePreview = previewImageModal.querySelector(".modal__title-preview");
-const altPreview = previewImageModal.querySelector("#preview-image-of-modal");
+// const previewImageModal = document.querySelector("#preview-image");
+// const imgPreview = previewImageModal.querySelector(
+//   ".modal__card-image-preview"
+// );
+// const titlePreview = previewImageModal.querySelector(".modal__title-preview");
+// const altPreview = previewImageModal.querySelector("#preview-image-of-modal");
 
 class Card {
-  constructor({ name, link }, cardSelector) {
-    this._name = name;
-    this._link = link;
-
+  constructor({ data, handleImageClick }, cardSelector) {
+    this._name = data.name;
+    this._link = data.link;
     this._cardSelector = cardSelector;
+    this._handleImageClick = handleImageClick;
   }
 
   _setEventListeners() {
@@ -28,10 +29,10 @@ class Card {
         this._handleTrashButton();
       });
 
-    this._element
+    this._cardImage
       .querySelector(".card__image")
       .addEventListener("click", () => {
-        this._handlePreviewPicture();
+        this._handleImageClick({ link: this._link, name: this._name });
       });
   }
 
@@ -45,27 +46,27 @@ class Card {
     this._element.remove();
     this._element = null;
   }
-  _handlePreviewPicture() {
-    imgPreview.src = this._link;
-    altPreview.alt = this._name;
-    titlePreview.textContent = this._name;
-    openModal(previewImageModal);
-  }
+  // _handlePreviewPicture() {
+  //   imgPreview.src = this._link;
+  //   altPreview.alt = this._name;
+  //   titlePreview.textContent = this._name;
+  //   openModal(previewImageModal);
+  // }
 
   _getTemplate() {
-    const cardElement = document
+    return document
       .querySelector(this._cardSelector)
       .content.querySelector(".card")
       .cloneNode(true);
-
-    return cardElement;
   }
 
   getView() {
     this._element = this._getTemplate();
-    this._element.querySelector(".card__image").src = this._link;
+    this._cardImage = this._element.querySelector(".card__image");
+    this._cardImage.src = this._link;
+    this._cardImage.alt = this._name;
     this._element.querySelector(".card__title").textContent = this._name;
-    this._element.querySelector(".card__image").alt = this._name;
+
     this._setEventListeners();
 
     return this._element;
