+++ src/components/PopupWithForm.js
@@ -1,11 +1,28 @@
 import Popup from "./Popup.js";
 
-class PopupWithForm extends Popup {
+export default class PopupWithForm extends Popup {
   constructor({ popupSelector, handleFormSubmit }) {
     super({ popupSelector });
     this._popupForm = this._popupElement.querySelector(".modal__form");
     this._handleFormSubmit = handleFormSubmit;
   }
+  _getInputValues() {
+    const inputList = this._popupForm.querySelectorAll(".modal__input");
+    const data = {};
+    inputList.forEach((input) => {
+      data[input.name] = input.value;
+    });
+    return data;
+  }
+
+  setEventListeners() {
+    super.setEventListeners();
+    this._popupForm.addEventListeners("submit", (evt) => {
+      evt.preventDefault();
+      this._handleFormSubmit(this._getInputValues());
+    });
+  }
+
   close() {
     this._popupForm.reset();
     super.close();
