+++ src/components/Section.js
@@ -18,6 +18,6 @@ export default class Section {
   }
 
   addItem(element) {
-    this._container(element);
+    this._container.append(element);
   }
 }
