+++ src/components/UserInfo.js
@@ -1,0 +1,18 @@
+export default class UserInfo {
+  constructor(titleSelector, descriptionSelector) {
+    this._titleElement = document.querySelector(titleSelector);
+    this._descriptionElement = doucment.querySelector(descriptionSelector);
+  }
+
+  getUserInfo() {
+    return {
+      title: this._titleElement.textContent,
+      description: this._descriptionElement.textContent,
+    };
+  }
+
+  setUserInfo(title, description) {
+    this._titleElement.textContent = title;
+    this._descriptionElement.textContent = description;
+  }
+}
